Move code to packages and rename
