+++ webapi/src/main/scala/org/knora/webapi/messages/util/ValueUtilV1.scala
@@ -14,12 +14,12 @@ import org.knora.webapi._
 import org.knora.webapi.config.AppConfig
 import org.knora.webapi.core.MessageRelay
 import org.knora.webapi.messages.OntologyConstants
-import org.knora.webapi.messages.StandoffStuff
 import org.knora.webapi.messages.StringFormatter
 import org.knora.webapi.messages.ValuesValidator
 import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
 import org.knora.webapi.messages.util.GroupedProps._
 import org.knora.webapi.messages.util.rdf.VariableResultsRow
+import org.knora.webapi.messages.util.standoff.StandoffStringUtil
 import org.knora.webapi.messages.util.standoff.StandoffTagUtilV2
 import org.knora.webapi.messages.v1.responder.ontologymessages._
 import org.knora.webapi.messages.v1.responder.resourcemessages.LiteralValueType
@@ -893,7 +893,7 @@ final case class ValueUtilV1Live(
       standoff = standoffTags,
       mappingIri = mappingIri,
       mapping = mappingResponse.mapping,
-      resource_reference = StandoffStuff.getResourceIrisFromStandoffTags(standoffTags)
+      resource_reference = StandoffStringUtil.getResourceIrisFromStandoffTags(standoffTags)
     )
 
   }
