+++ webapi/src/main/scala/org/knora/webapi/messages/v1/responder/valuemessages/ValueMessagesV1.scala
@@ -17,12 +17,12 @@ import org.knora.webapi._
 import org.knora.webapi.core.RelayedMessage
 import org.knora.webapi.messages.OntologyConstants
 import org.knora.webapi.messages.ResponderRequest.KnoraRequestV1
-import org.knora.webapi.messages.StandoffStuff
 import org.knora.webapi.messages.StringFormatter
 import org.knora.webapi.messages.admin.responder.projectsmessages.ProjectADM
 import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
 import org.knora.webapi.messages.traits.Jsonable
 import org.knora.webapi.messages.util.DateUtilV1
+import org.knora.webapi.messages.util.standoff.StandoffStringUtil
 import org.knora.webapi.messages.util.standoff.StandoffTagUtilV2
 import org.knora.webapi.messages.v1.responder.KnoraResponseV1
 import org.knora.webapi.messages.v1.responder.resourcemessages.LocationV1
@@ -714,7 +714,7 @@ case class TextValueWithStandoffV1(
       standoffNode: StandoffTagV2 =>
         CreateStandoffTagV1InTriplestore(
           standoffNode = standoffNode,
-          standoffTagInstanceIri = StandoffStuff.makeRandomStandoffTagIri(
+          standoffTagInstanceIri = StandoffStringUtil.makeRandomStandoffTagIri(
             valueIri = valueIri,
             startIndex = standoffNode.startIndex
           ) // generate IRI for new standoff node
