+++ webapi/src/main/scala/org/knora/webapi/messages/v2/responder/valuemessages/ValueMessagesV2.scala
@@ -23,7 +23,6 @@ import org.knora.webapi.messages.OntologyConstants
 import org.knora.webapi.messages.OntologyConstants.KnoraApiV2Complex._
 import org.knora.webapi.messages.ResponderRequest.KnoraRequestV2
 import org.knora.webapi.messages.SmartIri
-import org.knora.webapi.messages.StandoffStuff
 import org.knora.webapi.messages.StringFormatter
 import org.knora.webapi.messages.ValuesValidator
 import org.knora.webapi.messages.admin.responder.projectsmessages.ProjectADM
@@ -33,6 +32,7 @@ import org.knora.webapi.messages.store.sipimessages.GetFileMetadataResponse
 import org.knora.webapi.messages.util.PermissionUtilADM.EntityPermission
 import org.knora.webapi.messages.util._
 import org.knora.webapi.messages.util.rdf._
+import org.knora.webapi.messages.util.standoff.StandoffStringUtil
 import org.knora.webapi.messages.util.standoff.StandoffTagUtilV2
 import org.knora.webapi.messages.util.standoff.XMLUtil
 import org.knora.webapi.messages.v2.responder._
@@ -1549,7 +1549,7 @@ case class TextValueContentV2(
         standoffNode: StandoffTagV2 =>
           CreateStandoffTagV2InTriplestore(
             standoffNode = standoffNode,
-            standoffTagInstanceIri = StandoffStuff.makeRandomStandoffTagIri(
+            standoffTagInstanceIri = StandoffStringUtil.makeRandomStandoffTagIri(
               valueIri = valueIri,
               startIndex = standoffNode.startIndex
             ) // generate IRI for new standoff node
